factorized the NBT and data in the FuelHeater class. Regression to fix
