+++ src/main/resources/assets/tkcya/lang/en_us.lang
@@ -506,6 +506,8 @@ behavior.tricorder.max_heat=Maximum Heat: %s HU
 behavior.tricorder.burn_time_left= Burn Time Left %s ticks
 behavior.tricorder.currentTemp=Current Temp: %s K
 behavior.tricorder.heatMultiplier=Heat Multiplier: %d
+behavior.tricorder.can_ignite=Can ignite: %s
+behavior.tricorder.is_burning=Is burning: %s
 
 
 # Errors
